requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
requests
beautifulsoup4